+++ app/controllers/CountyCtrl.java
@@ -1,7 +1,7 @@
 package controllers;
 
 import credentials.CredentialsCheckerAction;
-import models.Category;
+import models.County;
 import org.myweb.db.Dao;
 import org.myweb.services.crud.*;
 import play.db.jpa.Transactional;
@@ -9,14 +9,14 @@ import play.mvc.Controller;
 import play.mvc.Result;
 import play.mvc.With;
 
-public class CategoryCtrl extends Controller {
+public class CountyCtrl extends Controller {
 
     @Transactional(readOnly = true)
     @With(CredentialsCheckerAction.class)
     public static Result get(Long id){
 
         return GetService.getInstance(Dao.getInstance())
-                .get( Category.class, id )
+                .get( County.class, id )
                 .buildPlayCtrlResult();
 
     }
@@ -26,7 +26,7 @@ public class CategoryCtrl extends Controller {
     public static Result query(){
 
         return QueryService.getInstance(Dao.getInstance())
-                .query(Category.class)
+                .query(County.class)
                 .buildPlayCtrlResult();
 
     }
@@ -36,7 +36,7 @@ public class CategoryCtrl extends Controller {
     public static Result update(Long id){
 
         return UpdateService.getInstance(Dao.getInstance())
-                .update( Category.class, request().body().asJson(), id )
+                .update( County.class, request().body().asJson(), id )
                 .buildPlayCtrlResult();
 
     }
@@ -46,7 +46,7 @@ public class CategoryCtrl extends Controller {
     public static Result create(){
 
         return CreateService.getInstance(Dao.getInstance())
-                .create( Category.class, request().body().asJson() )
+                .create( County.class, request().body().asJson() )
                 .buildPlayCtrlResult();
 
 
@@ -57,7 +57,7 @@ public class CategoryCtrl extends Controller {
     public static Result delete(Long id){
 
         return DeleteService.getInstance(Dao.getInstance())
-                .delete( Category.class, id )
+                .delete( County.class, id )
                 .buildPlayCtrlResult();
 
     }